requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# aria2rpc/__init__.py: 4
python_baseconv == 1.2.0

# aria2rpc/__init__.py: 2
requests == 2.18.4
